Approximate ZCTA-level ACS data with census tracts

Approximate ZCTA (ZIP Code Tabulation Area) level ACS variables using available census-tract level ACS data, for any year between 2009 and 2019. This process utilizes both the tidycensus package and a tract-to-ZCTA crosswalk.

acs_tract_to_zcta_fun(state_num, acs_vars, acs_year)

Arguments

state_num

A 2-digit state FIPS code. Both numeric and character values are accepted. ACS data can only be downloaded for one state at a time.

acs_vars

A vector of ACS 5-year survey variables.

acs_year

A numeric year which indicates the final year in the 5-year ACS survey data (e.g., acs_year = 2015 results in 2011-2015 ACS data). The value must be between 2009 and 2019.

Value

A data frame containing the best approximation of ZCTA-level ACS data for the state and year provided. The data will contain one row per combination of ZCTA & ACS variable. The data frame will contain the following columns:

state_fips - (character) 2-digit state FIPS code, matching the provided value of state_num.

zcta - (character) 5-digit ZCTA ID code

variable - (character) the ACS variable(s) requested.

est_zcta - (numeric) the ZCTA-level estimate of the ACS variable.

p_zcta_covered - (numeric) the proportion (valued 0-1) of the ZCTA (weighted by population) that was able to be matched with census tracts and, in essense, "covered" by the aggregated full and partial tract-level estimates. For example, a value of 0.5 indicates that only 50% of the ZCTA is represented by this approximated value.

Details

ACS 5-year survey variables generally begin with "B". Note that the tidycensus format for ACS variables includes an underscore between the table and variable number, e.g., "B01001_001". For a full list of ACS 5-year survey variable names, see tidycensus::load_variables().

This process only works for count data. Do not use this function for non-count variables (medians, proportions, etc.).

Data source used for ZCTA-tract crosswalks: Missouri Census Data Center: Geographic Correspondence Engine. Web page: https://mcdc.missouri.edu/applications/geocorr.html

If you are new to tidycensus, and need to set up an API key for census data, see this help page to get started: https://walker-data.com/tidycensus/articles/basic-usage.html

Examples

temp_pop_ny_2009 <- acs_tract_to_zcta_fun("36","B01001_001",2009)
#> [1] "using the 2000 tract-to-ZCTA crosswalk"
#> Getting data from the 2005-2009 5-year ACS
temp_pop_ny_pa_2009 <- do.call(rbind, lapply(c("36","42"),function(i){
  acs_tract_to_zcta_fun(i,"B01001_001",2009)}))
#> [1] "using the 2000 tract-to-ZCTA crosswalk"
#> Getting data from the 2005-2009 5-year ACS
#> [1] "using the 2000 tract-to-ZCTA crosswalk"
#> Getting data from the 2005-2009 5-year ACS