Determine county proxies for the largest cities in the US

Create a data set containing county proxy information for the largest cities in the US, according to custom city-county match criteria.

city_county_proxies_fun(n_cities = 50, year, rule, min_p_county, min_p_city)

Arguments

n_cities

The number of large cities for which county proxies are desired. Defaults to 50.

year

The year of population data used to rank city sizes and determine the largest n_cities cities.

rule

The rule used for determining whether a county should be considered an appropriate proxy for the city. Options are "or", "and", "county", and "city".

"county": Only the min_p_county requirement must be satisfied. min_p_city is not considered.

"city": Only the min_p_city requirement is considered. min_p_county is not considered.

"and": Both min_p_county and min_p_city requirements must be satisfied.

"or": Either the min_p_county requirement or min_p_city requirement must be satisfied.

min_p_county

The minimum proportion of the county that must overlap with the city (weighted by population) for the county to be considered an appropriate proxy for the city. Proportions are valued 0.0001 (0.01%) to 1 (100%).

min_p_city

The minimum proportion of the city that must overlap with the county (weighted by population) for the county to be considered an appropriate proxy for the city. Proportions are valued 0.0001 (0.01%) to 1 (100%).

Value

A data frame containing a crosswalk between cities and appropriate county proxies. The data frame will contain the following columns:

rank - (numeric) rank of the city by population size. The largest city has a rank of 1

city_name - (character) the name of the city or place

state_abbr (character) - the 2-letter state abbreviation for the city

city_fips - (character) the 7-digit FIPS code for the city, created by concatenating the 2-digit state FIPS code and 5-digit place FIPS code

county_name - (character) the county name (with state abbreviation) for the city's county proxy

county_fips - (character) the 5-digit FIPS code for the county, created by concatenating the 2-digit state FIPS code and 3-digit county FIPS code

p_city - (numeric) the proportion of the city that overlaps with the county (weighted by population), valued 0 - 1

p_county - (numeric) the proportion of the county that overlaps with the city (weighted by population), valued 0 - 1

city_pop_est - (numeric) the population estimate for the city in the year specified in the input

Details

For this data, cities are defined as places, which can be incorporated places (such as cities, towns, or villages), or census designated places (CDPs).

"Counties" includes several types of county-equivalent entities, such as:

Parishes (Louisiana)

Independent cities (Maryland, Missouri, Nevada, Virginia)

District of Columbia (Washington, DC)

Boroughs, municipalities, and census areas (Alaska)

Cities in the US often do not cleanly follow the boundaries of US counties. The following situations are possible for a city:

the city is contained wholly within one county, with a remaining portion of the county not belonging to that city;

the city is equivalent to one or more counties, e.g., the 5 boroughs of New York, NY which are all distinct counties;

the city overlaps with portions of multiple counties. In these cases, a portion of a county that overlaps with a certain city may potentially be very small, or vice versa.

Limitations of the Data

During the process to create the requested data, cities will be sorted by population size before applying the min_p_city and/or min_p_county rules. Some cities may be eliminated from the data entirely if no appropriate county matches are found which satisfy the user's conditions, especially for large values of min_p_city and/or min_p_county. If that happens, the number of cities in the final resulting crosswalk may be less than the requested n_cities.

Over time, some counties and cities/towns change names, split into multiple parts, or merge into other areas. In these cases, not only will the name of a county or place change, but so will its geographic ID/FIPS code.

Due to limitations of the source data, some cities which ceased to exist (in their original forms) prior to 2019 may not be included in the data at all, or may be listed as different (post-2019 equivalent) names or FIPS codes than what would have existed at an earlier time. User-generated crosswalks for earlier years may include some cities which did not exist yet in those earlier years.

Additionally, the county proxies returned in the requested crosswalk data may be different according to which year the user inputs. Requests for years 2010-2019 utilize a 2010 county-place relationship file, while requests for years 2020 onward utilize a 2020 county-place relationship file.

Data sources used:

American National Standards Institute (ANSI) and Federal Information Processing Series (FIPS) Codes: 2010 ANSI Codes for Places. Web page: https://www.census.gov/library/reference/code-lists/ansi.html#place

Relationship Files: State-level 2020 Place to 2010 Place Relationship Files. Web page: https://www.census.gov/geographies/reference-files/time-series/geo/relationship-files.2020.html#place

Missouri Census Data Center: Geographic Correspondence Engine. Web page: https://mcdc.missouri.edu/applications/geocorr.html

City and Town Population Totals: 2020-2021. Data set: Annual Estimates of the Resident Population for Incorporated Places of 50,000 or More, Ranked by July 1, 2021 Population: April 1, 2020 to July 1, 2021. Web page: https://www.census.gov/data/tables/time-series/demo/popest/2020s-total-cities-and-towns.html

City and Town Population Totals: 2010-2019 Data set: Annual Estimates of the Resident Population for Incorporated Places of 50,000 or More, Ranked by July 1, 2019 Population: April 1, 2010 to July 1, 2019. Web page: https://www.census.gov/data/tables/time-series/demo/popest/2020s-total-cities-and-towns.html

Examples

city_county_proxies_fun(n_cities=50, year=2020, rule="city", min_p_city=0.05)
#> [1] "NOTE: rule='city', so min_p_county is ignored."
#> [1] "City selection rule selected: at least 5% of city overlaps with county."
#>    rank                                              city_name state_abbr
#> 1     1                                          New York city         NY
#> 2     1                                          New York city         NY
#> 3     1                                          New York city         NY
#> 4     1                                          New York city         NY
#> 5     1                                          New York city         NY
#> 6     2                                       Los Angeles city         CA
#> 7     3                                           Chicago city         IL
#> 8     4                                           Houston city         TX
#> 9     5                                           Phoenix city         AZ
#> 10    6                                      Philadelphia city         PA
#> 11    7                                       San Antonio city         TX
#> 12    8                                         San Diego city         CA
#> 13    9                                            Dallas city         TX
#> 14   10                                          San Jose city         CA
#> 15   11                                            Austin city         TX
#> 16   11                                            Austin city         TX
#> 17   12                                      Jacksonville city         FL
#> 18   13                                        Fort Worth city         TX
#> 19   14                                          Columbus city         OH
#> 20   15                            Indianapolis city (balance)         IN
#> 21   16                                         Charlotte city         NC
#> 22   17                                     San Francisco city         CA
#> 23   18                                           Seattle city         WA
#> 24   19                                            Denver city         CO
#> 25   20                                        Washington city         DC
#> 26   21   Nashville-Davidson metropolitan government (balance)         TN
#> 27   22                                     Oklahoma City city         OK
#> 28   22                                     Oklahoma City city         OK
#> 29   22                                     Oklahoma City city         OK
#> 30   23                                           El Paso city         TX
#> 31   24                                            Boston city         MA
#> 32   25                                          Portland city         OR
#> 33   26                                         Las Vegas city         NV
#> 34   27                                           Detroit city         MI
#> 35   28 Louisville/Jefferson County metro government (balance)         KY
#> 36   29                                           Memphis city         TN
#> 37   30                                         Baltimore city         MD
#> 38   31                                         Milwaukee city         WI
#> 39   32                                       Albuquerque city         NM
#> 40   33                                            Fresno city         CA
#> 41   34                                            Tucson city         AZ
#> 42   35                                        Sacramento city         CA
#> 43   36                                       Kansas City city         MO
#> 44   36                                       Kansas City city         MO
#> 45   36                                       Kansas City city         MO
#> 46   37                                              Mesa city         AZ
#> 47   38                                           Atlanta city         GA
#> 48   38                                           Atlanta city         GA
#> 49   39                                             Omaha city         NE
#> 50   40                                  Colorado Springs city         CO
#> 51   41                                           Raleigh city         NC
#> 52   42                                        Long Beach city         CA
#> 53   43                                    Virginia Beach city         VA
#> 54   44                                             Miami city         FL
#> 55   45                                           Oakland city         CA
#> 56   46                                       Minneapolis city         MN
#> 57   47                                             Tulsa city         OK
#> 58   48                                       Bakersfield city         CA
#> 59   49                                           Wichita city         KS
#> 60   50                                         Arlington city         TX
#>    city_fips             county_name county_fips p_city p_county city_pop_est
#> 1    3651000             Richmond NY       36085 0.0563   1.0000      8772978
#> 2    3651000                Bronx NY       36005 0.1673   1.0000      8772978
#> 3    3651000             New York NY       36061 0.1924   1.0000      8772978
#> 4    3651000               Queens NY       36081 0.2732   1.0000      8772978
#> 5    3651000                Kings NY       36047 0.3108   1.0000      8772978
#> 6    0644000          Los Angeles CA       06037 1.0000   0.3893      3889834
#> 7    1714000                 Cook IL       17031 1.0000   0.5206      2741730
#> 8    4835000               Harris TX       48201 0.9787   0.4767      2300027
#> 9    0455000             Maricopa AZ       04013 1.0000   0.3638      1611345
#> 10   4260000         Philadelphia PA       42101 1.0000   1.0000      1601005
#> 11   4865000                Bexar TX       48029 1.0000   0.7140      1438227
#> 12   0666000            San Diego CA       06073 1.0000   0.4205      1385394
#> 13   4819000               Dallas TX       48113 0.9384   0.4684      1303234
#> 14   0668000          Santa Clara CA       06085 1.0000   0.5233      1010908
#> 15   4805000           Williamson TX       48491 0.0627   0.0990       963121
#> 16   4805000               Travis TX       48453 0.9364   0.6981       963121
#> 17   1235000                Duval FL       12031 1.0000   0.9538       950463
#> 18   4827000              Tarrant TX       48439 0.9801   0.4267       922592
#> 19   3918000             Franklin OH       39049 0.9719   0.6650       905860
#> 20   1836003               Marion IN       18097 1.0000   0.9083       887382
#> 21   3712000          Mecklenburg NC       37119 1.0000   0.7840       876747
#> 22   0667000        San Francisco CA       06075 1.0000   1.0000       870014
#> 23   5363000                 King WA       53033 1.0000   0.3247       738172
#> 24   0820000               Denver CO       08031 1.0000   1.0000       717630
#> 25   1150000 District of Columbia DC       11001 1.0000   1.0000       690093
#> 26   4752006             Davidson TN       47037 1.0000   0.9631       689248
#> 27   4055000            Cleveland OK       40027 0.1142   0.2632       682760
#> 28   4055000             Canadian OK       40017 0.1134   0.5002       682760
#> 29   4055000             Oklahoma OK       40109 0.7723   0.6605       682760
#> 30   4824000              El Paso TX       48141 1.0000   0.7842       678598
#> 31   2507000              Suffolk MA       25025 1.0000   0.8467       674272
#> 32   4159000            Multnomah OR       41051 0.9962   0.7972       652388
#> 33   3240000                Clark NV       32003 1.0000   0.2833       643292
#> 34   2622000                Wayne MI       26163 1.0000   0.3563       638176
#> 35   2148006            Jefferson KY       21111 1.0000   0.4941       632037
#> 36   4748000               Shelby TN       47157 1.0000   0.6809       631326
#> 37   2404000       Baltimore city MD       24510 1.0000   1.0000       583132
#> 38   5553000            Milwaukee WI       55079 1.0000   0.6144       576301
#> 39   3502000           Bernalillo NM       35001 1.0000   0.8346       564648
#> 40   0627000               Fresno CA       06019 1.0000   0.5375       542159
#> 41   0477000                 Pima AZ       04019 1.0000   0.5200       541859
#> 42   0664000           Sacramento CA       06067 1.0000   0.3312       523416
#> 43   2938000              Jackson MO       29095 0.6228   0.4412       507932
#> 44   2938000               Platte MO       29165 0.1050   0.4999       507932
#> 45   2938000                 Clay MO       29047 0.2720   0.5454       507932
#> 46   0446000             Maricopa AZ       04013 1.0000   0.1141       505860
#> 47   1304000               DeKalb GA       13089 0.0802   0.0524       499586
#> 48   1304000               Fulton GA       13121 0.9198   0.4300       499586
#> 49   3137000              Douglas NE       31055 1.0000   0.8315       491168
#> 50   0816000              El Paso CO       08041 1.0000   0.6558       480213
#> 51   3755000                 Wake NC       37183 0.9967   0.4127       467425
#> 52   0643000          Los Angeles CA       06037 1.0000   0.0466       464759
#> 53   5182000  Virginia Beach city VA       51810 1.0000   1.0000       459373
#> 54   1245000           Miami-Dade FL       12086 1.0000   0.1637       441889
#> 55   0653000              Alameda CA       06001 1.0000   0.2619       439341
#> 56   2743000             Hennepin MN       27053 1.0000   0.3355       429014
#> 57   4075000                Tulsa OK       40143 0.9857   0.6084       412629
#> 58   0603526                 Kern CA       06029 1.0000   0.4437       403401
#> 59   2079000             Sedgwick KS       20173 1.0000   0.7589       397117
#> 60   4804000              Tarrant TX       48439 1.0000   0.1868       393985
city_county_proxies_fun(n_cities=25, year=2010, rule="county", min_p_county=0.1)
#> [1] "NOTE: rule='county', so min_p_city is ignored."
#> [1] "County selection rule selected: at least 10% of county overlaps with city."
#>    rank                                            city_name state_abbr
#> 1     1                                        New York city         NY
#> 2     1                                        New York city         NY
#> 3     1                                        New York city         NY
#> 4     1                                        New York city         NY
#> 5     1                                        New York city         NY
#> 6     2                                     Los Angeles city         CA
#> 7     3                                         Chicago city         IL
#> 8     4                                         Houston city         TX
#> 9     5                                    Philadelphia city         PA
#> 10    6                                         Phoenix city         AZ
#> 11    7                                     San Antonio city         TX
#> 12    8                                       San Diego city         CA
#> 13    9                                          Dallas city         TX
#> 14   10                                        San Jose city         CA
#> 15   11                                    Jacksonville city         FL
#> 16   12                          Indianapolis city (balance)         IN
#> 17   13                                          Austin city         TX
#> 18   14                                   San Francisco city         CA
#> 19   15                                        Columbus city         OH
#> 20   16                                      Fort Worth city         TX
#> 21   17                                       Charlotte city         NC
#> 22   18                                         Detroit city         MI
#> 23   19                                         Memphis city         TN
#> 24   20                                         El Paso city         TX
#> 25   21                                          Boston city         MA
#> 26   22                                       Baltimore city         MD
#> 27   23                                         Seattle city         WA
#> 28   24                                      Washington city         DC
#> 29   25 Nashville-Davidson metropolitan government (balance)         TN
#>    city_fips             county_name county_fips p_city p_county city_pop_est
#> 1    3651000             Richmond NY       36085 0.0563   1.0000      8190209
#> 2    3651000                Bronx NY       36005 0.1673   1.0000      8190209
#> 3    3651000             New York NY       36061 0.1924   1.0000      8190209
#> 4    3651000               Queens NY       36081 0.2732   1.0000      8190209
#> 5    3651000                Kings NY       36047 0.3108   1.0000      8190209
#> 6    0644000          Los Angeles CA       06037 1.0000   0.3893      3795512
#> 7    1714000                 Cook IL       17031 1.0000   0.5206      2697477
#> 8    4835000               Harris TX       48201 0.9787   0.4767      2100280
#> 9    4260000         Philadelphia PA       42101 1.0000   1.0000      1528283
#> 10   0455000             Maricopa AZ       04013 1.0000   0.3638      1449038
#> 11   4865000                Bexar TX       48029 1.0000   0.7140      1332299
#> 12   0666000            San Diego CA       06073 1.0000   0.4205      1305906
#> 13   4819000               Dallas TX       48113 0.9384   0.4684      1200350
#> 14   0668000          Santa Clara CA       06085 1.0000   0.5233       954940
#> 15   1235000                Duval FL       12031 1.0000   0.9538       823114
#> 16   1836003               Marion IN       18097 1.0000   0.9083       821579
#> 17   4805000               Travis TX       48453 0.9364   0.6981       806164
#> 18   0667000        San Francisco CA       06075 1.0000   1.0000       805505
#> 19   3918000             Franklin OH       39049 0.9719   0.6650       790943
#> 20   4827000              Tarrant TX       48439 0.9801   0.4267       748441
#> 21   3712000          Mecklenburg NC       37119 1.0000   0.7840       738444
#> 22   2622000                Wayne MI       26163 1.0000   0.3563       711131
#> 23   4748000               Shelby TN       47157 1.0000   0.6809       652326
#> 24   4824000              El Paso TX       48141 1.0000   0.7842       650671
#> 25   2507000              Suffolk MA       25025 1.0000   0.8467       621048
#> 26   2404000       Baltimore city MD       24510 1.0000   1.0000       620915
#> 27   5363000                 King WA       53033 1.0000   0.3247       610630
#> 28   1150000 District of Columbia DC       11001 1.0000   1.0000       605226
#> 29   4752006             Davidson TN       47037 1.0000   0.9631       604589
city_county_proxies_fun(n_cities=60, year=2010, rule="and", min_p_county=0.025, min_p_city=0.15)
#> [1] "County selection rule selected: at least 2.5% of county overlaps with city, AND at least 15% of city overlaps with county."
#>    rank                                              city_name state_abbr
#> 1     1                                          New York city         NY
#> 2     1                                          New York city         NY
#> 3     1                                          New York city         NY
#> 4     1                                          New York city         NY
#> 5     2                                       Los Angeles city         CA
#> 6     3                                           Chicago city         IL
#> 7     4                                           Houston city         TX
#> 8     5                                      Philadelphia city         PA
#> 9     6                                           Phoenix city         AZ
#> 10    7                                       San Antonio city         TX
#> 11    8                                         San Diego city         CA
#> 12    9                                            Dallas city         TX
#> 13   10                                          San Jose city         CA
#> 14   11                                      Jacksonville city         FL
#> 15   12                            Indianapolis city (balance)         IN
#> 16   13                                            Austin city         TX
#> 17   14                                     San Francisco city         CA
#> 18   15                                          Columbus city         OH
#> 19   16                                        Fort Worth city         TX
#> 20   17                                         Charlotte city         NC
#> 21   18                                           Detroit city         MI
#> 22   19                                           Memphis city         TN
#> 23   20                                           El Paso city         TX
#> 24   21                                            Boston city         MA
#> 25   22                                         Baltimore city         MD
#> 26   23                                           Seattle city         WA
#> 27   24                                        Washington city         DC
#> 28   25   Nashville-Davidson metropolitan government (balance)         TN
#> 29   26                                            Denver city         CO
#> 30   27 Louisville/Jefferson County metro government (balance)         KY
#> 31   28                                         Milwaukee city         WI
#> 32   29                                          Portland city         OR
#> 33   30                                         Las Vegas city         NV
#> 34   31                                     Oklahoma City city         OK
#> 35   32                                       Albuquerque city         NM
#> 36   33                                            Tucson city         AZ
#> 37   34                                            Fresno city         CA
#> 38   35                                        Sacramento city         CA
#> 39   36                                        Long Beach city         CA
#> 40   37                                       Kansas City city         MO
#> 41   37                                       Kansas City city         MO
#> 42   38                                             Omaha city         NE
#> 43   39                                              Mesa city         AZ
#> 44   40                                    Virginia Beach city         VA
#> 45   41                                           Atlanta city         GA
#> 46   42                                  Colorado Springs city         CO
#> 47   43                                           Raleigh city         NC
#> 48   44                                             Miami city         FL
#> 49   45                                         Cleveland city         OH
#> 50   46                                             Tulsa city         OK
#> 51   47                                           Oakland city         CA
#> 52   48                                       Minneapolis city         MN
#> 53   49                                           Wichita city         KS
#> 54   50                                         Arlington city         TX
#> 55   51                                       Bakersfield city         CA
#> 56   52                                       New Orleans city         LA
#> 57   53                                     Urban Honolulu CDP         HI
#> 58   54                                             Tampa city         FL
#> 59   55                                           Anaheim city         CA
#> 60   56                                            Aurora city         CO
#> 61   57                                         Santa Ana city         CA
#> 62   58                                         St. Louis city         MO
#> 63   59                                         Riverside city         CA
#> 64   60                                        Pittsburgh city         PA
#>    city_fips             county_name county_fips p_city p_county city_pop_est
#> 1    3651000                Bronx NY       36005 0.1673   1.0000      8190209
#> 2    3651000             New York NY       36061 0.1924   1.0000      8190209
#> 3    3651000               Queens NY       36081 0.2732   1.0000      8190209
#> 4    3651000                Kings NY       36047 0.3108   1.0000      8190209
#> 5    0644000          Los Angeles CA       06037 1.0000   0.3893      3795512
#> 6    1714000                 Cook IL       17031 1.0000   0.5206      2697477
#> 7    4835000               Harris TX       48201 0.9787   0.4767      2100280
#> 8    4260000         Philadelphia PA       42101 1.0000   1.0000      1528283
#> 9    0455000             Maricopa AZ       04013 1.0000   0.3638      1449038
#> 10   4865000                Bexar TX       48029 1.0000   0.7140      1332299
#> 11   0666000            San Diego CA       06073 1.0000   0.4205      1305906
#> 12   4819000               Dallas TX       48113 0.9384   0.4684      1200350
#> 13   0668000          Santa Clara CA       06085 1.0000   0.5233       954940
#> 14   1235000                Duval FL       12031 1.0000   0.9538       823114
#> 15   1836003               Marion IN       18097 1.0000   0.9083       821579
#> 16   4805000               Travis TX       48453 0.9364   0.6981       806164
#> 17   0667000        San Francisco CA       06075 1.0000   1.0000       805505
#> 18   3918000             Franklin OH       39049 0.9719   0.6650       790943
#> 19   4827000              Tarrant TX       48439 0.9801   0.4267       748441
#> 20   3712000          Mecklenburg NC       37119 1.0000   0.7840       738444
#> 21   2622000                Wayne MI       26163 1.0000   0.3563       711131
#> 22   4748000               Shelby TN       47157 1.0000   0.6809       652326
#> 23   4824000              El Paso TX       48141 1.0000   0.7842       650671
#> 24   2507000              Suffolk MA       25025 1.0000   0.8467       621048
#> 25   2404000       Baltimore city MD       24510 1.0000   1.0000       620915
#> 26   5363000                 King WA       53033 1.0000   0.3247       610630
#> 27   1150000 District of Columbia DC       11001 1.0000   1.0000       605226
#> 28   4752006             Davidson TN       47037 1.0000   0.9631       604589
#> 29   0820000               Denver CO       08031 1.0000   1.0000       603359
#> 30   2148006            Jefferson KY       21111 1.0000   0.4941       596482
#> 31   5553000            Milwaukee WI       55079 1.0000   0.6144       594865
#> 32   4159000            Multnomah OR       41051 0.9962   0.7972       585429
#> 33   3240000                Clark NV       32003 1.0000   0.2833       584576
#> 34   4055000             Oklahoma OK       40109 0.7723   0.6605       582516
#> 35   3502000           Bernalillo NM       35001 1.0000   0.8346       546755
#> 36   0477000                 Pima AZ       04019 1.0000   0.5200       527160
#> 37   0627000               Fresno CA       06019 1.0000   0.5375       497769
#> 38   0664000           Sacramento CA       06067 1.0000   0.3312       467286
#> 39   0643000          Los Angeles CA       06037 1.0000   0.0466       462431
#> 40   2938000              Jackson MO       29095 0.6228   0.4412       460697
#> 41   2938000                 Clay MO       29047 0.2720   0.5454       460697
#> 42   3137000              Douglas NE       31055 1.0000   0.8315       459659
#> 43   0446000             Maricopa AZ       04013 1.0000   0.1141       440993
#> 44   5182000  Virginia Beach city VA       51810 1.0000   1.0000       438859
#> 45   1304000               Fulton GA       13121 0.9198   0.4300       429410
#> 46   0816000              El Paso CO       08041 1.0000   0.6558       420685
#> 47   3755000                 Wake NC       37183 0.9967   0.4127       406353
#> 48   1245000           Miami-Dade FL       12086 1.0000   0.1637       400779
#> 49   3916000             Cuyahoga OH       39035 1.0000   0.2946       395926
#> 50   4075000                Tulsa OK       40143 0.9857   0.6084       392403
#> 51   0653000              Alameda CA       06001 1.0000   0.2619       391406
#> 52   2743000             Hennepin MN       27053 1.0000   0.3355       383067
#> 53   2079000             Sedgwick KS       20173 1.0000   0.7589       382745
#> 54   4804000              Tarrant TX       48439 1.0000   0.1868       365892
#> 55   0603526                 Kern CA       06029 1.0000   0.4437       348987
#> 56   2255000       Orleans Parish LA       22071 1.0000   1.0000       347772
#> 57   1571550             Honolulu HI       15003 1.0000   0.3453       339121
#> 58   1271000         Hillsborough FL       12057 1.0000   0.2637       337053
#> 59   0602000               Orange CA       06059 1.0000   0.1088       336760
#> 60   0804000             Arapahoe CO       08005 0.8700   0.5130       326055
#> 61   0669000               Orange CA       06059 1.0000   0.0973       325327
#> 62   2965000       St. Louis city MO       29510 1.0000   1.0000       319336
#> 63   0662000            Riverside CA       06065 1.0000   0.1303       305466
#> 64   4261000            Allegheny PA       42003 1.0000   0.2423       305283