Create a ZIP Code-ZCTA crosswalk for a range of years.

Function to create a best-match ZIP Code-to-ZCTA crosswalk over a multi-year period.

zip_zcta_cw_fun(years = c(2009:2021), po_city_names = T)

Arguments

years

a numeric vector of years (limited to 2009 through 2021)

po_city_names

A Boolean (T/F) indicating whether to include post office city names and states of ZIP Codes in the returned data.

Value

A data.frame containing the best-match crosswalk between ZIP Code and ZCTA.

The crosswalk contains one row per ZIP Code. In many cases, several ZIP Codes may match to a single ZCTA.

Columns in the returned data are: "zip" (5 digit ZIP Code), "zcta" (5-digit ZCTA code), "po_city" (post office city name), "state" (2-letter state abbreviation). If po_city_names = F, then the two columns "po_city" and "state" will not be included.

Details

Over time, many details about ZIP Codes change, including geographical location and boundaries, post office names, population sizes, and type (standard, PO box, military, etc.). ZCTA (ZIP Code Tabulation Area) assignments for ZIP codes may also change by year.

The best ZCTA match for each ZIP Code over a span of multiple years is determined via algorithm using the following considerations, in order, to narrow down options:

(1) ZIP Code Type: ZIP Code type may change over time. Populated standard ZIP Codes have priority over other ZIP Code types, such as PO box;

(2) Consistency: The ZCTA assigned to the ZIP Code for the greatest number of years is selected;

(3) Time: The most recent year's ZIP-ZCTA match is selected out of the remaining options, if there are still multiple matches.

Post office city names are determined by a similar but simpler set of considerations:

(1) Consistency: The city name assigned to the ZIP Code for the greatest number of years is selected;

(2) Time: The most recent year's ZIP Code PO city name assignment is selected out of the remaining options, if there are still multiple matches.

Single-year ZIP-ZCTA tables are sourced from: https://udsmapper.org

Examples

cw_2009_2021 <- zip_zcta_cw_fun()
#> [1] 2009
#> [1] 2010
#> [1] 2011
#> [1] 2012
#> [1] 2013
#> [1] 2014
#> [1] 2015
#> [1] 2016
#> [1] 2017
#> [1] 2018
#> [1] 2019
#> [1] 2020
#> [1] 2021
#> [1] 41476
#> [1] 41476
#> [1] 41476
#> [1] 41476
#> [1] 41476
#> [1] 41476
#> [1] 41476
cw_2017_2021 <- zip_zcta_cw_fun(years = c(2017:2021))
#> [1] 2017
#> [1] 2018
#> [1] 2019
#> [1] 2020
#> [1] 2021
#> [1] 41180
#> [1] 41180
#> [1] 41180
#> [1] 41180
#> [1] 41180
#> [1] 41180
#> [1] 41180